requirements.txt
Django<5.0,>=4.0
djangorestframework<4.0,>=3.13
psycopg2-binary<3.0,>=2.9
pytest>=7.0,<8.0
pytest-django>=4.0,<5.0
djangorestframework-simplejwt>=5.2.2,<6.0
pylint>=2.0,<3.0
black
isort
